requirements.txt
Flask==3.0.0
openai==1.57.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
urllib3==2.1.0
httpx==0.25.2
limits==3.6.0